minor bug fix

+++ main/routes.py
@@ -377,7 +377,8 @@ def delete_message(message_id):
     message = Messages.query.filter_by(id = message_id).first_or_404()
     message_reply = MessageReply.query.filter_by(message_id = message_id).first()
     db.session.delete(message)
-    db.session.delete(message_reply)
+    if message_reply is not None:
+        db.session.delete(message_reply)
     db.session.commit()
     return redirect(url_for('view_messages'))
